EcoIND mc hist shows Ecopath baseline if computed MC shows avg #iterations needed to find balanced models Minor relabeling in Computed cap UI
git-svn-id: https://sources.ecopath.org/svn/Ecopath/trunk/Ecopath6@16150 a725721c-0ed6-864e-a6a8-02c2a1988c9d

+++ Sources/EwECore/Monte Carlo/cMonteCarloManager.vb
@@ -614,7 +614,7 @@ Public Class cMonteCarloManager
     ''' the current trial.
     ''' </summary>
     ''' -----------------------------------------------------------------------
-    Public ReadOnly Property nEcopathIterations() As Single
+    Public ReadOnly Property nEcopathIterations() As Integer
         Get
             If (Me.m_mc Is Nothing) Then Return 0
             Return m_mc.nEcopathIterations
